# news-subscriber/news-subscriber-bff/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Article API
    version: 0.0.1
paths:
    /newssub/v1/article/create:
        post:
            tags:
                - Article
            operationId: Article_CreateArticle
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.newssub.v1.CreateArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.newssub.v1.CreateArticleReply'
    /newssub/v1/article/{uid}:
        get:
            tags:
                - Article
            operationId: Article_GetArticle
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.newssub.v1.GetArticleReply'
        put:
            tags:
                - Article
            operationId: Article_UpdateArticle
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.newssub.v1.UpdateArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.newssub.v1.UpdateArticleReply'
        delete:
            tags:
                - Article
            operationId: Article_DeleteArticle
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.newssub.v1.DeleteArticleReply'
    /newssub/v1/list-article:
        get:
            tags:
                - Article
            operationId: Article_ListArticle
            parameters:
                - name: start
                  in: query
                  schema:
                    type: string
                - name: end
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.newssub.v1.ListArticleReply'
components:
    schemas:
        api.newssub.v1.ArticleItem:
            type: object
            properties:
                uid:
                    type: string
                title:
                    type: string
                content:
                    type: string
                tags:
                    type: array
                    items:
                        type: string
        api.newssub.v1.CreateArticleReply:
            type: object
            properties:
                uid:
                    type: string
        api.newssub.v1.CreateArticleRequest:
            type: object
            properties:
                title:
                    type: string
                content:
                    type: string
                tags:
                    type: array
                    items:
                        type: string
        api.newssub.v1.DeleteArticleReply:
            type: object
            properties: {}
        api.newssub.v1.GetArticleReply:
            type: object
            properties:
                article:
                    $ref: '#/components/schemas/api.newssub.v1.ArticleItem'
        api.newssub.v1.ListArticleReply:
            type: object
            properties:
                articles:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.newssub.v1.ArticleItem'
        api.newssub.v1.UpdateArticleReply:
            type: object
            properties: {}
        api.newssub.v1.UpdateArticleRequest:
            type: object
            properties:
                uid:
                    type: string
                title:
                    type: string
                content:
                    type: string
                tags:
                    type: array
                    items:
                        type: string
tags:
    - name: Article